Add unittest of ffmpeg worker.
