+++ api/utils/environment_variables.go
@@ -38,7 +38,6 @@ const (
 	EnvDisableFaceRecognition EnvironmentVariable = "PHOTOVIEW_DISABLE_FACE_RECOGNITION"
 	EnvDisableVideoEncoding   EnvironmentVariable = "PHOTOVIEW_DISABLE_VIDEO_ENCODING"
 	EnvDisableRawProcessing   EnvironmentVariable = "PHOTOVIEW_DISABLE_RAW_PROCESSING"
-	EnvVideoEncoder           EnvironmentVariable = "PHOTOVIEW_VIDEO_ENCODER"
 )
 
 // GetName returns the name of the environment variable itself
